refactor(code-editor): extract codeMirrorShared — theme + language stacks
The theme/highlight/language exports added to CodeMirrorEditor.tsx broke
react-refresh (component files must export only components). They move to
a non-component module consumed by all three viewers; the lazy-only gate
lists it.
